Fix Int32 sign-extension in make_group_key
Missed in the original Int32 cast fix sweep — hierarchical join group
keys for Int32 columns (parquet physical type for UInt32) were still
sign-extending via `as u64` instead of reinterpreting via `as u32`.

+++ src/join/hierarchical.rs
@@ -54,7 +54,7 @@ fn make_group_key(
             if a.is_null(row) {
                 return Ok(None);
             }
-            buf.extend_from_slice(&(a.value(row) as u64).to_le_bytes());
+            buf.extend_from_slice(&((a.value(row) as u32) as u64).to_le_bytes());
         } else {
             return Err(anyhow!(
                 "unsupported group key column type: {:?}",
